# Improve GDAL Performance on Integration Tests