--- a/gdal/src/it/scala/geotrellis/contrib/vlm/SubsceneReadingIT.scala
+++ b/gdal/src/it/scala/geotrellis/contrib/vlm/SubsceneReadingIT.scala
@@ -27,7 +27,10 @@ import org.scalatest.{FunSpec, Matchers}
 class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
   import Timing._
 
-  val sample = "https://s3-us-west-2.amazonaws.com/radiant-nasa-iserv/2014/02/14/IP0201402141023382027S03100E/IP0201402141023382027S03100E-COG.tif"
+  val sample = "http://s3-us-west-2.amazonaws.com/radiant-nasa-iserv/2014/02/14/IP0201402141023382027S03100E/IP0201402141023382027S03100E-COG.tif"
+  // val sample = "s3://radiant-nasa-iserv/2014/02/14/IP0201402141023382027S03100E/IP0201402141023382027S03100E-COG.tif"
+  val sample2 = "s3://radiant-nasa-iserv/2014/02/14/IP0201402141023382027S03100E/IP0201402141023382027S03100E-COG.tif"
+
   implicit class WithSubExtent(e: Extent) {
     /** Constructs an arbitrary subextent covering 1% of base extent. */
     def subextent: Extent = {
@@ -42,6 +45,7 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
   val readCells = (r: Raster[MultibandTile]) => r.mapTile(_.mapDouble((_, c) => c)).size
 
   describe("GDAL vs GeoTrellis GeoTiff reading") {
+
     it("should read projected extent with similar performance with GDAL") {
       val gt = time("GeoTrellis") {
         val src = GeoTiffRasterSource(sample)
@@ -50,7 +54,7 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
       logger.info(gt.toString)
 
       val gdal = time("GDAL") {
-        val src = GDALRasterSource(sample)
+        val src = GDALRasterSource(sample2)
         ProjectedExtent(src.extent, src.crs)
       }
       logger.info(gdal.toString)
@@ -58,6 +62,7 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
       gt.result should be (gdal.result)
       gt.durationMillis should be(gdal.durationMillis +- 1000)
     }
+
     it("should be faster reading subextent with GDAL") {
       val gt = time("GeoTrellis") {
         val src = GeoTiffRasterSource(sample)
@@ -66,13 +71,14 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
       logger.info(gt.toString)
 
       val gdal = time("GDAL") {
-        val src = GDALRasterSource(sample)
+        val src = GDALRasterSource(sample2)
         src.read(src.extent.subextent).map(readCells)
       }
       logger.info(gdal.toString)
 
       gt.durationMillis shouldBe >= (gdal.durationMillis)
     }
+
     it("should be faster reading full scene as tiles with GDAL") {
       def subbounds(src: RasterSource) = src.gridBounds.split(256, 256).toArray
       val gt = time("GeoTrellis") {
@@ -84,7 +90,7 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
       logger.info(gt.toString)
 
       val gdal = time("GDAL") {
-        val src = GDALRasterSource(sample)
+        val src = GDALRasterSource(sample2)
         val bnds = subbounds(src)
         bnds.flatMap(b => src.read(b, Seq(0))).map(readCells).sum
       }
@@ -101,7 +107,7 @@ class SubsceneReadingIT extends FunSpec with Matchers with LazyLogging {
       logger.info(gt.toString)
 
       val gdal = time("GDAL") {
-        val src = GDALRasterSource(sample)
+        val src = GDALRasterSource(sample2)
         src.read(Seq(0)).map(readCells).get
       }
       logger.info(gdal.toString)