@pomadchin 2019-04-04
https://github.com/OSGeo/gdal/pull/1417 by @jamesmcclain :tada: