@jamesmcclain 2019-04-04
I am now getting these numbers:
```
[info] SubsceneReadingIT:
[info] GDAL vs GeoTrellis GeoTiff reading
09:22:05 SubsceneReadingIT: Elapsed time of GeoTrellis is 1.7920s, producing ProjectedExtent(Extent(3440101.0175691093, -2315767.899836476, 3462532.0983280675, -2294035.527693799),EPSG:3857)
09:22:06 SubsceneReadingIT: Elapsed time of GDAL is 1.0770s, producing ProjectedExtent(Extent(3440101.0175691093, -2315767.899836476, 3462532.0983280675, -2294035.527693799),merc-CS)
[info] - should be faster reading projected extent with GDAL (3 seconds, 7 milliseconds)
09:22:08 SubsceneReadingIT: Elapsed time of GeoTrellis is 2.3870s, producing Some(372620)
09:22:09 SubsceneReadingIT: Elapsed time of GDAL is 1.2670s, producing Some(372620)
[info] - should be faster reading subextent with GDAL (3 seconds, 656 milliseconds)
09:22:10 SubsceneReadingIT: reading 600 tiles
09:22:48 SubsceneReadingIT: Elapsed time of GeoTrellis is 38.7800s, producing 37194588
09:22:56 SubsceneReadingIT: Elapsed time of GDAL is 8.0180s, producing 37194588
[info] - should be faster reading full scene as tiles with GDAL (46 seconds, 801 milliseconds)
09:23:38 SubsceneReadingIT: Elapsed time of GeoTrellis is 41.7260s, producing 37194588
09:23:38 SubsceneReadingIT: Elapsed time of GDAL is 0.1100s, producing 37194588
[info] - should be faster reading full scene GDAL (41 seconds, 837 milliseconds)
[info] Run completed in 1 minute, 37 seconds.
[info] Total number of tests run: 4
[info] Suites: completed 1, aborted 0
[info] Tests: succeeded 4, failed 0, canceled 0, ignored 0, pending 0
[info] All tests passed.
[success] Total time: 124 s, completed Apr 4, 2019, 9:23:38 AM
```

The read performance looks really good (8 seconds then 0 seconds [caching]) -- this might be because I am now using GDAL master instead of 2.4.1, or it might be that the configuration options are really just that good.