@jamesmcclain 2019-04-04
> [OSGeo/gdal#1417](https://github.com/OSGeo/gdal/pull/1417) by @jamesmcclain

GDAL PR closed, but solution found :+1: